Splitting sanity-check stuff into separate file to be included for now, removed when testing is done.
